+++ src/Import.hs
@@ -6,8 +6,10 @@ module Import (
   module Types,
   module Data.Complex,
   module Util,
+  module Codec.Picture,
 ) where
 
+import Codec.Picture
 import Data.Complex
 import RIO
 import RIO.File