+++ src/Run.hs
@@ -9,12 +9,11 @@ import Import
 run :: RIO App ()
 run = do
   logInfo "Calculating inputs..."
-  !app <- ask
-  let !opts = appOptions app
-      !graphBounds = optionsInputBounds opts
+  !opts <- asks appOptions
+  let !graphBounds = optionsInputBounds opts
       !res = optionsGraphPts opts
-      !output = serializeImage $ graph fn graphBounds res
-      !outFile = optionsOutFile opts
+  !output <- serializeImage <$> graph fn graphBounds res
+  let !outFile = optionsOutFile opts
   logInfo $ "Completed calculations. Writing output to " <> display outFile <> "."
   writeBinaryFile outFile output
 
@@ -22,3 +21,15 @@ run = do
 -- so long as it typechecks
 fn :: (RealFloat a) => Complex a -> Complex a
 fn = id
+
+graph :: (Complex Double -> Complex Double) -> Bounds -> Int -> RIO App DynamicImage
+graph !f !bnds !res = do
+  !clr <- asks (not . optionsDontUseColor . appOptions)
+  !blk <- asks (not . optionsDontUseBlack . appOptions)
+  !wht <- asks (not . optionsDontUseWhite . appOptions)
+  let calcH = if clr then getH else const 240
+      calcS = if wht then getS else const 1
+      calcV = if blk then getV else const 1
+      complexToHsv !z = {-# SCC complexToHsv #-} (calcH z, calcS z, calcV z)
+      createPixel !a !b = hsvToRgb . complexToHsv . f $ genInput bnds res a b
+  return $! ImageRGBF $ generateImage createPixel res res