+++ src/Util.hs
@@ -5,11 +5,12 @@
 -- | Utils for complex-hsv.
 module Util (
   genInput,
-  graph,
   serializeImage,
   clamp,
-  -- exported for testing:
-  complexToHsv,
+  hsvToRgb,
+  getH,
+  getS,
+  getV,
 ) where
 
 import Codec.Picture
@@ -28,11 +29,6 @@ genInput (!na, !xa, !nb, !xb) !res !ia !ib = (na + a * da) :+ (nb + b * db)
   !a = fromIntegral ia
   !b = fromIntegral ib
 
-graph :: (Complex Double -> Complex Double) -> Bounds -> Int -> DynamicImage
-graph !fn !bnds !res = ImageRGBF $ generateImage createPixel res res
- where
-  createPixel !a !b = {-# SCC createPixel #-} hsvToRgb . complexToHsv . fn $ genInput bnds res a b
-
 hsvToRgb :: (PixelF, PixelF, PixelF) -> PixelRGBF
 hsvToRgb (!h, !s, !v) = toPixel $ hsv h s v
 
@@ -42,14 +38,20 @@ toPixel (RGB !r !g !b) = PixelRGBF r g b
 serializeImage :: DynamicImage -> ByteString
 serializeImage = L.toStrict . encodePng . convertRGB8
 
-complexToHsv :: (RealFloat a) => Complex a -> (Float, Float, Float)
-complexToHsv !z = (realToFrac adjustedPhase, realToFrac . clamp 0 1 $ 1 / magnitude z, 1)
+getH :: (RealFloat a) => Complex a -> Float
+getH !z = realToFrac adjustedPhase
  where
   !adjustedPhase =
     if straightPhase < 0
       then straightPhase + 360
       else straightPhase
   !straightPhase = phase z / pi * 180
+
+getS :: (RealFloat a) => Complex a -> Float
+getS !z = realToFrac . clamp 0 1 $! recip $! magnitude z
+
+getV :: (RealFloat a) => Complex a -> Float
+getV !z = realToFrac . clamp 0 1 $! magnitude z
 
 clamp :: (Ord a) => a -> a -> a -> a
 clamp !mn !mx = max mn . min mx